+++ client/src/testing.rs
@@ -1,5 +1,4 @@
 pub mod accessors;
-pub mod forwarding_utils;
 pub mod init;
 pub mod mocks;
