+++ server/tests/test_expired_session.rs
@@ -1,8 +1,8 @@
-mod helpers;
+mod common;
 
 use std::time::Duration;
 
-use helpers::{check_broadcast, check_forwarding, spawn_receive_for_client, Mode};
+use common::{check_broadcast, check_forwarding, spawn_receive_for_client, Mode};
 use ya_relay_client::{ClientBuilder, FailFast};
 use ya_relay_core::crypto::FallbackCryptoProvider;
 use ya_relay_core::utils::to_udp_url;
